Fixed Font issue
FIrefox follows access rules for @font-face. This makes it so that it will not use fonts from another domain (site.com vs store.site.com) unless you set a config option for the server. This is the Access-Control-Allow-Origin setting. This setting is enabled for all .eot, .ttf etc files.

The font used had an extension of .TTF which is not .ttf and so the Access-Control-Allow-Origin was not set.

Fixes #88